telemetry_worker_ngrok/requirements.txt
# Web framework
Flask==3.0.0
gunicorn==21.2.0

# GCP libraries
google-cloud-trace==1.11.0
google-cloud-logging==3.8.0
google-cloud-pubsub==2.18.0
google-auth==2.24.0

# OpenTelemetry
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-proto==1.21.0

# HTTP client
requests==2.31.0

# Caching (for deduplication)
redis==5.0.1

# Protobuf
protobuf==4.25.0

# Utilities
python-dotenv==1.0.0
